+++ src/app/routes/admin/routes/users/components/admin-user-one/admin-user-one.component.ts
@@ -4,7 +4,12 @@ import { inOutAnimation } from '@core/animations/enter-leave.animations';
 import { routesConfig } from '@core/config';
 import { MessageI } from '@interfaces';
 import { User } from '@models';
-import { ToastService, UserService } from '@services';
+import {
+  ModalService,
+  ToastService,
+  UserService,
+  ValidationsFormService,
+} from '@services';
 import { FullImageService } from '@shared/services/ui/full-image/full-image.service';
 import { TOAST_STATE } from '@shared/services/ui/toast/toast.service';
 
@@ -25,7 +30,9 @@ export class AdminUserOneComponent {
     private userService: UserService,
     private fullImage: FullImageService,
     private toastService: ToastService,
-    private router: Router
+    private router: Router,
+    private modal: ModalService,
+    private validationsFormService: ValidationsFormService
   ) {}
 
   showImage(image: string) {
@@ -33,20 +40,41 @@ export class AdminUserOneComponent {
   }
 
   onSubmit() {
-    const observable = this.user._id
-      ? this.userService.update(this.user)
-      : this.userService.create(this.user);
-    observable.subscribe({
-      next: (response) => this.onSuccess(response),
-      error: (error) => this.toastService.showToast(TOAST_STATE.error, error),
-    });
+    const validation = this.validationsFormService.validation(
+      'user',
+      this.user
+    );
+    if (validation.state) {
+      const observable = this.user._id
+        ? this.userService.update(this.user)
+        : this.userService.create(this.user);
+      observable.subscribe({
+        next: (response) => this.onSuccess(response),
+        error: (error) => this.toastService.showToast(TOAST_STATE.error, error),
+      });
+    } else {
+      this.toastService.showToast(TOAST_STATE.error, validation.message);
+    }
   }
 
   onDelete() {
-    // TODO: Añadir confirmacion por modal
-    this.userService.deleteOne(this.user._id!).subscribe({
-      next: (response) => this.onSuccess(response),
-      error: (error) => this.toastService.showToast(TOAST_STATE.error, error),
+    const modal = this.modal.showModalConfirm(
+      'Eliminar usuario',
+      '¿Estas seguro de eliminar el usuario?'
+    );
+    const sub$ = modal.subscribe({
+      next: (response) => {
+        if (response !== '') {
+          if (response === true) {
+            this.userService.deleteOne(this.user._id!).subscribe({
+              next: (response) => this.onSuccess(response),
+              error: (error) =>
+                this.toastService.showToast(TOAST_STATE.error, error),
+            });
+          }
+          sub$.unsubscribe();
+        }
+      },
     });
   }
